move scheduler diagnostics to diagnostics/
