+++ distributed/diagnostics/__init__.py
@@ -5,3 +5,5 @@ with ignoring(ImportError):
     from .progressbar import progress
 with ignoring(ImportError):
     from .resource_monitor import Occupancy
+with ignoring(ImportError):
+    from .scheduler_widgets import scheduler_status
